requirements.txt
fastapi
uvicorn[standard]
requests
python-dotenv
sentence-transformers
scikit-learn
numpy
sqlalchemy
psycopg2-binary
alembic
torch
scipy
